+++ R/autosuggest.R
@@ -21,7 +21,7 @@
 autosuggest <- function(address, results = 5, url_only = FALSE) {
 
   # Check addresses
-  .check_addresses(address)
+  .check_character(address)
   .check_numeric_range(results, 1, 100)
   .check_boolean(url_only)
 
@@ -34,7 +34,7 @@ autosuggest <- function(address, results = 5, url_only = FALSE) {
   url <- paste0(
     url,
     "&q=",
-    address
+    curl::curl_escape(address)
   )
 
   # Add bbox containing the world
@@ -56,8 +56,9 @@ autosuggest <- function(address, results = 5, url_only = FALSE) {
   }
 
   # Request and get content
-  data <- .get_content(
-    url = url
+  data <- .async_request(
+    url = url,
+    rps = 5
   )
   if (length(data) == 0) {
     return(NULL)