+++ R/geocode.R
@@ -36,7 +36,7 @@
 geocode <- function(address, alternatives = FALSE, sf = TRUE, url_only = FALSE) {
 
   # Input checks
-  .check_addresses(address)
+  .check_character(address)
   .check_boolean(alternatives)
   .check_boolean(sf)
   .check_boolean(url_only)
@@ -50,7 +50,7 @@ geocode <- function(address, alternatives = FALSE, sf = TRUE, url_only = FALSE) 
   url <- paste0(
     url,
     "&q=",
-    gsub("\\|", "", address)
+    curl::curl_escape(address)
   )
 
   # Return urls if chosen
@@ -59,8 +59,9 @@ geocode <- function(address, alternatives = FALSE, sf = TRUE, url_only = FALSE) 
   }
 
   # Request and get content
-  data <- .get_content(
-    url = url
+  data <- .async_request(
+    url = url,
+    rps = 5
   )
   if (length(data) == 0) {
     return(NULL)