+++ R/incident.R
@@ -70,8 +70,9 @@ incident <- function(aoi, from = Sys.time() - 60 * 60 * 24 * 7, url_only = FALSE
   }
 
   # Request and get content
-  data <- .get_content(
-    url = url
+  data <- .async_request(
+    url = url,
+    rps = Inf
   )
   if (length(data) == 0) {
     return(NULL)