+++ R/isoline.R
@@ -15,11 +15,10 @@
 #' @param routing_mode character, set the routing mode: \code{"fast"} or \code{"short"}.
 #' @param transport_mode character, set the transport mode: \code{"car"}, \code{"pedestrian"} or \code{"truck"}.
 #' @param traffic boolean, use real-time traffic or prediction in routing (\code{default = TRUE})? If no traffic is selected, the \code{datetime} is set to \code{"any"} and the request is processed independently from time.
+#' @param optimize, character, specifies how isoline calculation is optimized: \code{"balanced"}, \code{"quality"} or \code{"performance"} (\code{default = "balanced"}).
 #' @param consumption_model character, specify the consumption model of the vehicle, see \href{https://developer.here.com/documentation/routing-api/8.16.0/dev_guide/topics/use-cases/consumption-model.html}{consumption model} for more information (\code{default = NULL} a average electric car is set).
 #' @param aggregate boolean, aggregate (with function \code{min}) and intersect the isolines from geometry type \code{POLYGON} to geometry type \code{MULTIPOLYGON} (\code{default = TRUE})?
 #' @param url_only boolean, only return the generated URLs (\code{default = FALSE})?
-#' @param type character, 'type' is deprecated, use 'routing_mode' instead.
-#' @param mode character, 'mode' is deprecated, use 'transport_mode' instead.
 #'
 #' @return
 #' An \code{sf} object containing the requested isolines.
@@ -38,18 +37,9 @@
 isoline <- function(poi, datetime = Sys.time(), arrival = FALSE,
                     range = seq(5, 30, 5) * 60, range_type = "time",
                     routing_mode = "fast", transport_mode = "car",
-                    traffic = TRUE, consumption_model = NULL,
-                    aggregate = TRUE, url_only = FALSE, type, mode) {
-
-  # Deprecated parameters
-  if (!missing("type")) {
-    warning("'type' is deprecated, use 'routing_mode' instead.")
-    routing_mode <- type
-  }
-  if (!missing("mode")) {
-    warning("'mode' is deprecated, use 'transport_mode' instead.")
-    transport_mode <- mode
-  }
+                    traffic = TRUE, optimize = "balanced",
+                    consumption_model = NULL, aggregate = TRUE,
+                    url_only = FALSE) {
 
   # Checks
   .check_points(poi)
@@ -57,6 +47,7 @@ isoline <- function(poi, datetime = Sys.time(), arrival = FALSE,
   .check_range_type(range_type)
   .check_routing_mode(routing_mode)
   .check_transport_mode(transport_mode, request = "isoline")
+  .check_optimize(optimize)
   .check_boolean(traffic)
   .check_boolean(arrival)
   .check_boolean(aggregate)
@@ -111,6 +102,13 @@ isoline <- function(poi, datetime = Sys.time(), arrival = FALSE,
     range_type
   )
 
+  # Add optimization method
+  url <- paste0(
+    url,
+    "&optimizeFor=",
+    optimize
+  )
+
   # Add consumption model if specified, otherwise set to default electric vehicle
   if (is.null(consumption_model)) {
     url <- paste0(
@@ -141,8 +139,9 @@ isoline <- function(poi, datetime = Sys.time(), arrival = FALSE,
   }
 
   # Request and get content
-  data <- .get_content(
-    url = url
+  data <- .async_request(
+    url = url,
+    rps = 1
   )
   if (length(data) == 0) {
     return(NULL)
@@ -180,8 +179,7 @@ isoline <- function(poi, datetime = Sys.time(), arrival = FALSE,
   isolines <- sf::st_as_sf(
     isolines,
     sf_column_name = "geometry",
-    crs = 4326,
-    check_ring_dir = TRUE
+    crs = 4326
   )
 
   # Spatially aggregate
@@ -220,7 +218,18 @@ isoline <- function(poi, datetime = Sys.time(), arrival = FALSE,
           departure = if (arrival) NA else df$departure$time,
           arrival = if (arrival) df$arrival$time else NA,
           range = df$isolines$range$value,
-          geometry = sapply(df$isolines$polygons, function(x) x$outer)
+          geometry = lapply(df$isolines$polygons, function(x) {
+            # Decode flexible polyline encoding to ...
+            if (length(x$outer) > 1) {
+              # MULTIPOLYGON
+              sf::st_multipolygon(
+                sf::st_geometry(flexpolyline::decode_sf(x$outer, 4326))
+              )
+            } else {
+              # POLYGON
+              sf::st_geometry(flexpolyline::decode_sf(x$outer, 4326))[[1]]
+            }
+          })
         )
       })
     ),
@@ -232,11 +241,6 @@ isoline <- function(poi, datetime = Sys.time(), arrival = FALSE,
     return(NULL)
   }
 
-  # Decode flexible polyline encoding to POLYGON
-  geometry <- NULL
-  isolines[, "geometry" := sf::st_geometry(
-    flexpolyline::decode_sf(geometry, 4326)
-  )]
   return(isolines)
 }