+++ R/reverse_geocode.R
@@ -73,8 +73,9 @@ reverse_geocode <- function(poi, results = 1, sf = TRUE, url_only = FALSE) {
   }
 
   # Request and get content
-  data <- .get_content(
-    url = url
+  data <- .async_request(
+    url = url,
+    rps = 5
   )
   if (length(data) == 0) {
     return(NULL)