+++ tests/testthat/test-connection.R
@@ -22,7 +22,7 @@ test_that("connection works", {
   ## Test with API response mock
   # Route segments: "summary = FALSE"
   with_mock(
-    "hereR:::.get_content" = function(url) {
+    "hereR:::.async_request" = function(url, rps) {
       hereR:::mock$connection_response
     },
     connections <- connection(origin = poi[3:4, ], destination = poi[5:6, ], summary = FALSE),
@@ -34,7 +34,7 @@ test_that("connection works", {
 
   # Route summary: "summary = FALSE"
   with_mock(
-    "hereR:::.get_content" = function(url) {
+    "hereR:::.async_request" = function(url, rps) {
       hereR:::mock$connection_response
     },
     connections <- connection(origin = poi[3:4, ], destination = poi[5:6, ], summary = TRUE),