+++ NEWS.md
@@ -1,3 +1,13 @@
+# version 0.7.0
+
+* Enable `optimize` parameter to chose from "balanced", "quality" and "performance" in `isoline()` (closes [#119](https://github.com/munterfinger/hereR/issues/119)).
+* Remove deprecated parameters in `route()`, `route_matrix()` and `isoline()`.
+* Add specific user-agent to the requests: `hereR/<version> R/<version> (<platform>)`.
+* Add option to deactivate rate limits `set_rate_limit(FALSE)`.
+* Bugfix: Add rate limits in RPS (requests per seconds) to async requests to the APIs; move dependency for requests from **curl** to **crul** package (closes [#122](https://github.com/munterfinger/hereR/issues/122)).
+* Bugfix: `isoline()` now handles multipart polygons (MULTIPOLYGON) if received by the API (closes [#121](https://github.com/munterfinger/hereR/issues/121)).
+* Add area and feature avoidance in `route()` (closes [#117](https://github.com/munterfinger/hereR/issues/117)).
+
 # version 0.6.1
 
 * Use **styler** package and use `tyler::tidyverse_style()`to format the package.